Initial Setup
The challenge is to convert the C++ port of the Fluid HTN into GDExtension. This requires binding the methods to GdScript API. Which may or may not require reworking the C++ code to fit the needs of the problem space. GDExtension is more mature than GDNative but still there exists some confusion and short documentation. A wonderful template to get setup is provided and used. To actually implement the code will require conversion between C++ and Godot types using the C++ bindings. Each class needs to register its methods and parameters. F.Y.I. Support does exist for Signals.