Still attempting to determine calling GDScript functions
I have found that you can pass a node into the RefCounted and that gives you access to the Node and its functions. The code successfully passes the node into the Domain builder and sets up the operator. But it still seems to fail after the planner tick. The GDScript function exists but the object.hpp throws error when trying to call 'Object.call_internal'.