It calls MoveTo but returns null
Separating the HTN code into its own 3D node and then referencing the Agent's CharacterBody3D seems to solve all issues I was having. But now it doesn't seem to be actually calling the moveTo function in gdscript. This is confusing since it doesn't throw any errors but fails to print the message and return the bool from gdscript.